Add files via upload
I define an array named cars and put these items into it: toyota, honda, mazda, mitsubishi, tesla, lexus. Then logged the first and last item into console. For the last item i used the length property (-1). Then added an item mercedes-benz to the array  and log it to the console. Finally Looped through all items of the array and log them to the console for display.